Add FullScreenLayout component with props and use it in the [id] page
